moved some documentation to a new "retired" folder
