+++ Documentation/retired/BezierApproximation.docx
