+++ Documentation/retired/ClothoidSketch.pdf
