+++ Documentation/retired/IfcAlignmentChecker_HughWoods.xlsm
