+++ Documentation/retired/IfcAlignment_FormatSpecification.docx
